fix: align StaticOwoSdk with the real OWO 2.4.2 API
Verified against the SDK assembly in lib/net6.0/OWO.dll:

* OWO has no IsConnected boolean property; check
  ConnectionState == ConnectionState.Connected instead.
* OWO.Connect takes String[] not a single string; wrap the manual IP
  in a one-element array.
* OWO.Send is a two-arg method (Sensation, Muscle[]) — there is no
  single-arg overload that pulls muscles out of a SensationWithMuscles.
  Pass the resolved Muscle[] explicitly.
* SensationsFactory.Create takes int frequency, float duration, int
  intensity, float ramp_up, float ramp_down, float exit_delay
  (positionally) — switch from named arguments (which exposed us to
  parameter-name churn) to positional dispatch and propagate the int
  types through OwoSendCommand.

OwoBackend.BuildSendCommand now rounds frequency/intensity to int at
the boundary so StaticOwoSdk doesn't have to narrow. The intensity-
scaling unit test asserts on the int directly. dotnet build succeeds
both natively on Mac and as a Windows cross-build
(-p:OS=Windows_NT -p:EnableWindowsTargeting=true).

+++ src/Smited.Daemon.Abstractions/Backends/IOwoSdk.cs
@@ -70,15 +70,18 @@ public interface IOwoSdk
 /// <c>StaticOwoSdk</c> so this record stays free of SDK types.
 /// </summary>
 /// <remarks>
-/// All time fields are <strong>seconds</strong>. The OWO SDK's
-/// <c>SensationsFactory.Create(...)</c> takes seconds for both ramps,
-/// despite some older docs labelling them as milliseconds — the
-/// implementation here matches the v2.4.x signature.
+/// Field types match the OWO SDK 2.4.x signature for
+/// <c>SensationsFactory.Create</c>: <c>frequency</c> and
+/// <c>intensity</c> are <c>int</c>; durations are <c>float</c> seconds.
+/// (The SDK parameter for ramps is named <c>rampUpMillis</c> in the
+/// reflection metadata but the real-world callers pass second values
+/// like <c>0.1f</c>; treat the name as an SDK misnomer and document
+/// the units here.)
 /// </remarks>
 public sealed record OwoSendCommand(
-    float FrequencyHz,
+    int FrequencyHz,
     float DurationSeconds,
-    float IntensityPercentage,
+    int IntensityPercentage,
     float RampUpSeconds,
     float RampDownSeconds,
     float ExitDelaySeconds,
